move AddToCart out of controller namespace into DTO namespace
