+++ src/CoreShop/Bundle/OrderBundle/Factory/AddToCartFactory.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\OrderBundle\Factory;
 
-use CoreShop\Bundle\OrderBundle\Controller\AddToCart;
+use CoreShop\Bundle\OrderBundle\DTO\AddToCart;
 use CoreShop\Component\Order\Model\CartInterface;
 use CoreShop\Component\Order\Model\PurchasableInterface;
