+++ src/CoreShop/Bundle/OrderBundle/Factory/AddToCartFactoryInterface.php
@@ -12,7 +12,7 @@
 
 namespace CoreShop\Bundle\OrderBundle\Factory;
 
-use CoreShop\Bundle\OrderBundle\Controller\AddToCartInterface;
+use CoreShop\Bundle\OrderBundle\DTO\AddToCartInterface;
 use CoreShop\Component\Order\Model\CartInterface;
 use CoreShop\Component\Order\Model\PurchasableInterface;
